+++ dkron/assets_ui/generate.go
@@ -1,0 +1,3 @@
+package assets_ui
+
+//go:generate vfsgendev -source="github.com/distribworks/dkron/dkron/assets_ui".Assets