+++ dkron/job.go
@@ -50,6 +50,9 @@ var (
 
 // Job descibes a scheduled Job.
 type Job struct {
+	// Job id. Must be unique, it's a copy of name.
+	ID string `json:"id"`
+
 	// Job name. Must be unique, acts as the id.
 	Name string `json:"name"`
 
@@ -129,6 +132,7 @@ func NewJobFromProto(in *proto.Job) *Job {
 	next, _ := ptypes.Timestamp(in.GetNext())
 
 	job := &Job{
+		ID:             in.Name,
 		Name:           in.Name,
 		DisplayName:    in.Displayname,
 		Timezone:       in.Timezone,