+++ ui/src/types.ts
@@ -1,0 +1,7 @@
+import { ReduxState } from 'react-admin';
+
+export type ThemeName = 'light' | 'dark';
+
+export interface AppState extends ReduxState {
+    theme: ThemeName;
+}